Added dateCreated to masterArray
Added dateCreated to masterArray to properly keep track of the time each new mock is created. The former functionality has an error that generates the date every time the content is loaded from the local storage. This is inaccurate as the date displayed is the date of viewing and not the date in which the mock was created.